Oranização classe de extract

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ExtractRepository.java
@@ -3,5 +3,5 @@ package br.com.alelo.consumer.consumerpat.respository;
 import br.com.alelo.consumer.consumerpat.entity.Extract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExtractRepository extends JpaRepository<Extract, Integer> {
+public interface ExtractRepository extends JpaRepository<Extract, Long> {
 }
